no accurate stops

+++ source/carMoves.h
@@ -24,8 +24,8 @@ using namespace std;
 int checkVehicle(vehicle *veh1, vehicle *veh2);
 vector<vehicle *> carsInRoadNode(vector<vehicle *> vIengine, roadNode road);
 int myNewLaneIs(Position cPos, roadNode *nextRoad,void *arguments);
-int calcDistance(vector<int> Path, Position p1, Position p2, void *arguments);
-void *nextObstacle(vehicle *cv, bool mylane, int &dist, int &retType, void *arguments);
+double calcDistance(vector<int> Path, Position p1, Position p2, void *arguments);
+void *nextObstacle(vehicle *cv, bool mylane, double &dist, int &retType, void *arguments);
 bool carFits(vehicle *v, vector<vehicle *> vIengine,vector<roadNode> allRoads,void *arguments);
 bool canIovertake(vehicle *v, void *nextObs, int distanceFRONT, void *arguments);
